docs(story): asset documentpanel in new format
