+++ src/components/AssetDocumentsPanel/stories/basic.story.mdx
@@ -1,0 +1,89 @@
+import { Meta, Story, Preview, Props, Source } from '@storybook/addon-docs/blocks';
+import { AssetDocumentsPanel } from '../AssetDocumentsPanel';
+import { decorators } from './helper';
+
+<Meta title="Assets|AssetDocumentsPanel" decorators={decorators} />
+
+ # Asset Documents Panel
+
+This component loads list of documents related to the asset and presents it in the list of categories.
+Each item in the list can be expanded showing list of files.
+The component is used as one of the panes in `AssetMeta` component.
+
+**NOTE:** The component should have `ClientSDKProvider` as a parent component in react component tree.
+
+## Basic usage:
+
+Don't forget to import:
+
+```jsx
+import 'antd/dist/antd.css';
+
+import React from 'react';
+import { AssetDocumentsPanel } from '@cognite/gearbox';
+```
+
+<Preview>
+  <Story name="Basic Usage">
+    <AssetDocumentsPanel assetId={4650652196144007} />
+  </Story>
+</Preview>
+
+## Available props
+
+<Props of={AssetDocumentsPanel} />
+
+## Types
+<br/>
+<h4>FileRequestFilter</h4>
+
+This type can be imported from `@cognite/sdk`:
+
+```jsx
+import { FileRequestFilter } from `@cognite/sdk`;
+```
+
+#### OnDocumentClick
+
+Definition:
+
+```jsx
+type OnDocumentClick = (
+  document: Document,
+  category: string,
+  description: string
+) => void;
+```
+
+#### CollapseProps
+This type can be imported from `antd`:
+
+```jsx
+import { CollapseProps } from 'antd/lib/collapse';
+```
+
+#### DocumentRenderer
+
+Definition:
+
+```jsx
+import { FilesMetadata } from `@cognite/sdk`;
+
+type DocumentRenderer = (
+  document: FilesMetadata,
+  i: number,
+  documents: FilesMetadata[]
+) => React.ReactNode;
+
+```
+
+#### AssetDocumentsPanelStyles
+
+This interface defines inline CSS styles for inner elements of `AssetDocumentsPanel` component.
+The type can be imported from `@cognite/gearbox`:
+
+```jsx
+import { AssetDocumentsPanelStyles } from '@cognite/gearbox';
+```
+
+See more details in AssetMeta component documentation.
